+++ CHANGELOG.md
@@ -29,4 +29,6 @@
 * `strtools.filename` and `pathtools.parse_path` can now also work on
   `java.io.File` objects, which happen to be the type when using ImageJ2's
   *Script Parameter* `@# File`.
+* The dict returned by `pathtools.parse_path` now also contains the key
+  `basename` that provides the filename without extension.
 * Many improvements / clarifications in function docstrings.